+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [unreleased](https://github.com/uzh/pool/tree/HEAD)
 
+### Fixed
+
+- create operator via root UI
+
 ## [v0.0.1-beta.2](https://github.com/uzh/pool/tree/v0.0.1-beta.1) - 2022-12-13
 
 - release to test workflow using Github Workflow